+++ dessia_common/utils/types.py
@@ -6,9 +6,11 @@ Types tools
 """
 from typing import Any, Dict, List, Tuple, Type, Union, get_origin, get_args
 
-import json
+# import json
 from collections.abc import Iterator, Sequence
 from importlib import import_module
+
+import orjson
 
 import dessia_common as dc
 from dessia_common.typings import Subclass, InstanceOf, MethodType, ClassMethodType
@@ -51,15 +53,24 @@ def is_classname_transform(string: str):
     return False
 
 
-def is_jsonable(x):
+def is_jsonable(obj):
     """
     returns if object can be dumped as it is in a json
     """
+
+    # First trying with orjson which is more efficient
     try:
-        json.dumps(x)
+        orjson.dumps(obj, option=orjson.OPT_SERIALIZE_NUMPY).decode('utf-8')
         return True
-    except TypeError:
+    except Exception:
         return False
+
+    # # If an error occurs try with json
+    # try:
+    #     json.dumps(obj)
+    #     return True
+    # except TypeError:
+    #     return False
 
 
 def is_serializable(obj):